fix: 음성 녹음 api & 데이터 타입 수정

+++ domain/src/main/java/mago/apps/domain/model/answer/AnswerVoice.kt
@@ -2,7 +2,7 @@ package mago.apps.domain.model.answer
 
 data class AnswerVoice(
     val text: String,
-    val duration: Int,
+    val duration: Float,
     val voiceUrl: String,
     val waveformUrl: String
 )
